moved DirectoryRefreshModal
